requirements.txt
Flask==2.2.5
gunicorn==21.2.0
cloudinary==1.41.0
psycopg[binary]==3.2.12
google-generativeai>=0.3.0